+++ extension-manifest-v3/src/background/index.js
@@ -14,4 +14,4 @@ import './tab-stats.js';
 import './cosmetics.js';
 import './trackers-preview.js';
 import './communication.js';
-import './telemetry/index.js';
+import './telemetry.js';
