+++ extension-manifest-v3/src/background/telemetry.js
@@ -2,23 +2,42 @@ import { Metrics as Telemetry } from '@ghostery/libs';
 import { store } from 'hybrids';
 
 import Options from '/store/options.js';
-import UTMs from '/store/utms.js';
-
-import { loadStorage, saveStorage } from './storage';
 
 const log = console.log.bind(console, '[telemetry]');
 
 async function recordUTMs(telemetry, JUST_INSTALLED) {
   if (JUST_INSTALLED) {
     const { utm_source, utm_campaign } = await telemetry.detectUTMs();
-    await store.set(utms, { utm_source, utm_campaign });
+    // persist campaign & source only
+    await chrome.storage.local.set({ utms: { utm_campaign, utm_source } });
     return;
   }
 
-  const utms = await store.resolve(store.get(UTMs));
+  const { utms = {} } = await chrome.storage.local.get(['utms']);
 
   telemetry.setUTMs(utms);
 }
+
+const saveStorage = async (storage, metrics) => {
+  Object.assign(storage, metrics);
+  await chrome.storage.local.set({ metrics: storage });
+};
+
+const loadStorage = async () => {
+  const storage = {
+    active_daily_velocity: [],
+    engaged_daily_velocity: [],
+    engaged_daily_count: [],
+  };
+  Telemetry.FREQUENCY_TYPES.forEach((frequency) => {
+    Telemetry.CRITICAL_TYPES.forEach((type) => {
+      storage[`${type}_${frequency}`] = 0;
+    });
+  });
+  const { metrics = {} } = await chrome.storage.local.get(['metrics']);
+  Object.assign(storage, metrics);
+  return storage;
+};
 
 (async () => {
   const options = await store.resolve(store.get(Options));
